GhaDescriptor now supports different buffer types
