+++ clove/components/core/graphics/include/Clove/Graphics/Vulkan/VulkanDescriptor.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "Clove/Graphics/Descriptor.hpp"
+
+#include <vulkan/vulkan.h>
+
+namespace garlic::clove{
+    VkDescriptorType getDescriptorType(DescriptorType garlicType);
+}
